requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg[binary]==3.2.13
drf-yasg==1.21.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
django-storages==1.14.2
boto3==1.34.0
Pillow==10.4.0
requests==2.31.0